Move ShellApplication up a level to allow SB to auto-discover configurations.
