+++ spring-cloud-dataflow-shell/src/main/java/org/springframework/cloud/dataflow/shell/ShellApplication.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.dataflow.shell.app;
+package org.springframework.cloud.dataflow.shell;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
